requirements.txt
# requirements.txt
line-bot-sdk
Flask
gunicorn